Channel iteration rewrites

Begin to eliminate the uses of TrackList::Channels outside Track.cpp and WaveTrack.cpp,
including all except one (Repeat) in src/effects.

Instead use WaveTrack::Channels which return objects of the new WaveChannel class,
which does not inherit Track but has a sufficient interface for many purposes, including
getting and setting and appending of sample data, and queries of minimum, maximum,
and RMS values, which are used by effects and analyzers and display.

+++ libraries/lib-wave-track/WaveClip.cpp
@@ -887,7 +887,7 @@ void WaveClip::SetRate(int rate)
 {
    const auto trimLeftSampleNum = TimeToSamples(mTrimLeft);
    const auto trimRightSampleNum = TimeToSamples(mTrimRight);
-   auto ratio = mRate / rate;
+   auto ratio = static_cast<double>(mRate) / rate;
    mRate = rate;
    mTrimLeft = SamplesToTime(trimLeftSampleNum);
    mTrimRight = SamplesToTime(trimRightSampleNum);
